requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
nltk==3.8.1
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
dnspython==2.6.1